ExchangeRate inserted to DB

+++ src/main/java/junior/test/task/dto/ExchangeRateDto.java
@@ -3,13 +3,11 @@ package junior.test.task.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 public class ExchangeRateDto {
   private  Long id;
-  private BigDecimal kztValue;
-  private BigDecimal rubValue;
+  private double kztValue;
+  private double rubValue;
   private String timeLastUpdateUtc;
 }
