+++ src/main/java/junior/test/task/model/ExchangeRate.java
@@ -4,9 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
-
 @Entity
 @Table(name = "t_exchange_rates")
 @Getter
@@ -15,7 +12,7 @@ public class ExchangeRate {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
-  private BigDecimal kztValue;
-  private BigDecimal rubValue;
+  private double kztValue;
+  private double rubValue;
   private String timeLastUpdateUtc;
 }
